# use include_str! so that our macro is re-run if the given file changes