Before this change:

```
typify/example-macro $ cargo build
    Finished dev [unoptimized + debuginfo] target(s) in 0.03s
typify/example-macro $ touch ../example.json 
typify/example-macro $ cargo build -v
       Fresh unicode-xid v0.2.2
       Fresh cfg-if v1.0.0
       Fresh ucd-trie v0.1.3
       Fresh ppv-lite86 v0.2.10
       Fresh regex-syntax v0.6.25
       Fresh same-file v1.0.6
       Fresh lazy_static v1.4.0
       Fresh remove_dir_all v0.5.3
       Fresh home v0.5.3
       Fresh itoa v0.4.8
       Fresh dyn-clone v1.0.4
       Fresh convert_case v0.4.0
       Fresh pest v2.1.3
       Fresh walkdir v2.3.2
       Fresh proc-macro2 v1.0.29
       Fresh libc v0.2.103
       Fresh memchr v2.4.1
       Fresh semver-parser v0.10.2
       Fresh ryu v1.0.5
       Fresh quote v1.0.10
       Fresh getrandom v0.2.3
       Fresh aho-corasick v0.7.18
       Fresh semver v0.11.0
       Fresh syn v1.0.80
       Fresh rand_core v0.6.3
       Fresh regex v1.5.4
       Fresh serde_derive v1.0.130
       Fresh rand_chacha v0.3.1
       Fresh serde_derive_internals v0.25.0
       Fresh thiserror-impl v1.0.29
       Fresh toolchain_find v0.2.0
       Fresh serde v1.0.130
       Fresh rand v0.8.4
       Fresh thiserror v1.0.29
       Fresh schemars_derive v0.8.6
       Fresh tempfile v3.2.0
       Fresh serde_json v1.0.69
       Fresh rustfmt-wrapper v0.1.0
       Fresh schemars v0.8.6
       Fresh typify-impl v0.0.5-dev (/Users/ahl/src/typify/typify-impl)
       Fresh typify-macro v0.0.5-dev (/Users/ahl/src/typify/typify-macro)
       Fresh typify v0.0.5-dev (/Users/ahl/src/typify/typify)
       Fresh example-macro v0.0.5-dev (/Users/ahl/src/typify/example-macro)
    Finished dev [unoptimized + debuginfo] target(s) in 0.03s
```

With this change:

```
typify/example-macro $ cargo build
   Compiling typify-macro v0.0.5-dev (/Users/ahl/src/typify/typify-macro)
   Compiling typify v0.0.5-dev (/Users/ahl/src/typify/typify)
   Compiling example-macro v0.0.5-dev (/Users/ahl/src/typify/example-macro)
    Finished dev [unoptimized + debuginfo] target(s) in 2.15s
typify/example-macro $ touch ../example.json 
typify/example-macro $ cargo build -v
       Fresh unicode-xid v0.2.2
       Fresh cfg-if v1.0.0
       Fresh ucd-trie v0.1.3
       Fresh ppv-lite86 v0.2.10
       Fresh regex-syntax v0.6.25
       Fresh same-file v1.0.6
       Fresh itoa v0.4.8
       Fresh home v0.5.3
       Fresh remove_dir_all v0.5.3
       Fresh lazy_static v1.4.0
       Fresh dyn-clone v1.0.4
       Fresh convert_case v0.4.0
       Fresh pest v2.1.3
       Fresh walkdir v2.3.2
       Fresh proc-macro2 v1.0.29
       Fresh libc v0.2.103
       Fresh memchr v2.4.1
       Fresh semver-parser v0.10.2
       Fresh ryu v1.0.5
       Fresh quote v1.0.10
       Fresh getrandom v0.2.3
       Fresh aho-corasick v0.7.18
       Fresh semver v0.11.0
       Fresh syn v1.0.80
       Fresh rand_core v0.6.3
       Fresh regex v1.5.4
       Fresh serde_derive v1.0.130
       Fresh rand_chacha v0.3.1
       Fresh serde_derive_internals v0.25.0
       Fresh thiserror-impl v1.0.29
       Fresh toolchain_find v0.2.0
       Fresh rand v0.8.4
       Fresh serde v1.0.130
       Fresh thiserror v1.0.29
       Fresh schemars_derive v0.8.6
       Fresh serde_json v1.0.69
       Fresh tempfile v3.2.0
       Fresh rustfmt-wrapper v0.1.0
       Fresh schemars v0.8.6
       Fresh typify-impl v0.0.5-dev (/Users/ahl/src/typify/typify-impl)
       Fresh typify-macro v0.0.5-dev (/Users/ahl/src/typify/typify-macro)
       Fresh typify v0.0.5-dev (/Users/ahl/src/typify/typify)
   Compiling example-macro v0.0.5-dev (/Users/ahl/src/typify/example-macro)
     Running `rustc --crate-name example_macro --edition=2018 example-macro/src/main.rs --error-format=json --json=diagnostic-rendered-ansi --crate-type bin --emit=dep-info,link -C embed-bitcode=no -C split-debuginfo=unpacked -C debuginfo=2 -C metadata=81845111506e15e6 -C extra-filename=-81845111506e15e6 --out-dir /Users/ahl/src/typify/target/debug/deps -C incremental=/Users/ahl/src/typify/target/debug/incremental -L dependency=/Users/ahl/src/typify/target/debug/deps --extern serde=/Users/ahl/src/typify/target/debug/deps/libserde-f5383c985ff1b755.rlib --extern serde_json=/Users/ahl/src/typify/target/debug/deps/libserde_json-e73f3ce77e225189.rlib --extern typify=/Users/ahl/src/typify/target/debug/deps/libtypify-ad08c65677181d17.rlib`
    Finished dev [unoptimized + debuginfo] target(s) in 0.39s
```